+++ CHANGELOG.md
@@ -17,6 +17,7 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 - `SamDataView` now contains a view of the optional fields
 - Added `CigletIteratorChecked` to provide a fallible iterator over `Cigar`
 - Added `to_ciglet_iter_checked` to `ToCigletIterator`, providing fallible iterators
+- Added `op_consumes_query` and `op_consumes_ref` methods to `Ciglet`
 
 ### Changed